+++ engines/scumm/scumm.h
@@ -460,7 +460,7 @@ protected:
 	int getTalkSpeed();
 
 	// Scumm main loop & helper functions.
-	virtual void scummLoop(int delta);
+	virtual void scummLoop(byte delta);
 	virtual void scummLoop_updateScummVars();
 	virtual void scummLoop_handleSaveLoad();
 	virtual void scummLoop_handleDrawing();
@@ -476,7 +476,7 @@ public:
 protected:
 	virtual void parseEvent(Common::Event event);
 
-	void waitForTimer(int delay);
+	void waitForTimer(uint16 delay);
 	double getTimerFrequency() const;
 	double _msecFractionalParts;
 	uint32 _lastWaitTime;
